updated to fix an issue with time
